requirements.txt
psycopg2-binary==2.9.9
flask==3.0.2
python_dotenv==1.0.1
flask_sqlalchemy==3.1.1